# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Learn Platform API
  version: 0.1.0
  description: Self-hosted learning platform API
servers:
  - url: http://localhost:3001
    description: Development server
paths:
...
  /api/v1/me/avatar:
    put:
      summary: Replace the signed-in account's avatar with an uploaded image
      description: >-
        The request body is the RAW IMAGE BYTES, not multipart. There is one
        file and no other fields, so a multipart envelope would add a parser
        for untrusted input to a route whose whole purpose is to minimise how
        much untrusted parsing happens.

        The image is never stored as sent (design §11.1: "always re-encode,
        never serve the bytes you were given"). It is re-encoded to a 256x256
        WebP with all metadata dropped — a photograph uploaded straight from a
        phone carries the camera model and often a GPS fix, and publishing an
        avatar must not publish where it was taken.

        Accepted: image/jpeg, image/png, image/webp, decided by the file's own
        magic bytes rather than by this header or a filename. Everything else
        is refused, SVG emphatically included.
      operationId: putMyAvatar
      tags:
        - profile
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
          image/png:
            schema:
              type: string
              format: binary
          image/webp:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: The stored avatar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileAvatar'
        '400':
          description: >-
            Not one of the accepted formats, or not decodable as the format
            its header claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: >-
            The actor may not set an avatar. Anonymous callers, and operator
            accounts, which have no learner profile (§5.1).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: >-
            Larger than 2 MiB on the wire, or declaring more pixels than the
            decoder will be asked to allocate.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Remove the uploaded avatar, reverting to the generated identicon
      description: >-
        Idempotent: an account that never uploaded one still gets 204. There
        is no "no avatar" state — the identicon is always there underneath.
      operationId: deleteMyAvatar
      tags:
        - profile
      responses:
        '204':
          description: The account is back to its identicon
        '403':
          description: The actor may not change this avatar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/profiles/{handle}/avatar:
    get:
      summary: The image bytes of a learner's uploaded avatar
      description: >-
        Serves the stored WebP. Reachable without a session, exactly like the
        profile it belongs to, and rate-limited by the same per-IP limiter for
        the same reason — a face is part of the public profile of §11.

        404, not a placeholder, when the account has no uploaded avatar: the
        identicon is generated client-side from the seed in the profile
        payload, so there is nothing to serve here. 404 is also the answer for
        a handle that does not exist, so this endpoint is not an oracle for
        which handles are taken.
      operationId: getProfileAvatar
      tags:
        - profile
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
        - name: v
          in: query
          required: false
          description: >-
            Cache-busting token; the avatar's sha256, as published in the
            profile payload's `url`. Ignored by the server, which serves the
            current image whatever is passed.
          schema:
            type: string
      responses:
        '200':
          description: The avatar image
          content:
            image/webp:
              schema:
                type: string
                format: binary
        '304':
          description: Unchanged since the ETag the client holds
        '404':
          description: No such profile, or it has no uploaded avatar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests from this address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/me/profile:
    get:
      summary: The actor's own profile settings
      description: >-
        The account holder's view of their §11 profile page: bio, the
        `noindex` toggle, and the visibility of all five sections. A section
        the actor has never touched reads `private` here, because
        `profile_section_visibility` has no row for it and an absent row IS
        private (migration 0014) — there is no backfill.
      operationId: getMyProfileSettings
      tags:
        - profile
      responses:
        '200':
          description: The actor's profile settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileSettings'
        '403':
          description: >-
            The actor has no learner profile — an operator or teacher-only
            account (design §5, §5.1).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: The actor has no users row
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update the actor's own profile settings
      description: >-
        Sets any of bio, `noindex`, and per-section visibility. Fields left
        out are unchanged; sections left out of `visibility` are unchanged.
        An empty or blank bio stores NULL rather than an empty string.
      operationId: updateMyProfileSettings
      tags:
        - profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileSettingsUpdateRequest'
      responses:
        '200':
          description: The updated settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileSettings'
        '400':
          description: >-
            An unknown section name, an unknown visibility value, a
            non-boolean `noindex`, or a bio over 2000 characters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: The actor may not change this profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: The actor has no users row
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/profiles/{handle}:
    get:
      summary: A learner's profile page (design §11)
      description: >-
        The data behind `/u/{handle}`, built from the VIEWER's identity.

        Three viewers, three payloads. The owner sees every section
        regardless of its setting, plus their own settings. Another
        signed-in account sees the sections set to `signed_in` or `public`.
        An unauthenticated reader is served by a separate deny-by-default
        serializer with an explicit field allowlist, and sees only the
        sections set to `public`.

        A hidden section is ABSENT from `sections` — never sent for the
        client to hide. The email address is absent for every viewer,
        always. An unauthenticated feed entry carries no lesson slug,
        because §12 keeps lesson content behind login.

        Rate-limited per IP (the §13 limiter, reused). Reachable without a
        session, which is what `noindex` and the Open Graph tags on the web
        page are for.
      operationId: getProfile
      tags:
        - profile
      parameters:
        - name: handle
          in: path
          required: true
          description: The student-chosen handle (never derived from an email, §11).
          schema:
            type: string
      responses:
        '200':
          description: The profile, filtered to this viewer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '404':
          description: >-
            No such handle, or the account has no learner profile (§5.1:
            operator accounts have none).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many profile requests from this address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      description: An error response
      properties:
        message:
          type: string
          description: A human-readable error message
      required:
        - message
...
    SectionVisibility:
      type: string
      description: >-
        How far a section reaches. `private` is the default everywhere and is
        also what an absent `profile_section_visibility` row means.
      enum:
        - private
        - signed_in
        - public
    ProfileVisibility:
      type: object
      description: The visibility of all five sections. Every key is always present.
      properties:
        badges:
          $ref: '#/components/schemas/SectionVisibility'
        degrees:
          $ref: '#/components/schemas/SectionVisibility'
        courses:
          $ref: '#/components/schemas/SectionVisibility'
        activity_feed:
          $ref: '#/components/schemas/SectionVisibility'
        activity_heatmap:
          $ref: '#/components/schemas/SectionVisibility'
      required:
        - badges
        - degrees
        - courses
        - activity_feed
        - activity_heatmap
    ProfileSettings:
      type: object
      description: The account holder's own view of their profile configuration.
      properties:
        handle:
          type: string
        displayName:
          type:
            - string
            - "null"
          nullable: true
        bio:
          type:
            - string
            - "null"
          nullable: true
          description: Free text, at most 2000 characters. Blank is stored as null.
        noindex:
          type: boolean
          description: >-
            When true (the default for every account), the profile page emits
            `<meta name="robots" content="noindex, nofollow">`.
        avatar:
          $ref: '#/components/schemas/ProfileAvatar'
        visibility:
          $ref: '#/components/schemas/ProfileVisibility'
      required:
        - handle
        - displayName
        - bio
        - noindex
        - avatar
        - visibility
    ProfileSettingsUpdateRequest:
      type: object
      description: >-
        A partial update. Omitted fields are unchanged; sections omitted from
        `visibility` are unchanged.
      properties:
        bio:
          type:
            - string
            - "null"
          nullable: true
        noindex:
          type: boolean
        visibility:
          type: object
          description: Section → visibility, for the sections being changed.
          additionalProperties:
            $ref: '#/components/schemas/SectionVisibility'
    ProfileViewer:
      type: string
      description: >-
        Who this payload was built for, relative to its subject. Decided by
        the API from the session, never sent by the client.
      enum:
        - owner
        - signed_in
        - anonymous
    ProfileAvatar:
      type: object
      description: >-
        Which face to draw.

        `identicon` is the default and the fallback: `seed` is a one-way hash
        of the user id, so the face is stable without publishing the row key,
        and the client draws it with no request at all.

        `upload` means the account has a stored image; `url` points at
        `/api/v1/profiles/{handle}/avatar` with the content digest as a
        cache-busting query. `seed` is still present for `upload`, so a client
        whose image request fails has something to draw instead of a hole.
      properties:
        kind:
          type: string
          enum:
            - identicon
            - upload
        seed:
          type: string
        url:
          type:
            - string
            - "null"
          nullable: true
          description: Present and non-null only when `kind` is `upload`.
      required:
        - kind
        - seed
...
    ProfileSections:
      type: object
      description: >-
        The sections this viewer may see. A hidden section is ABSENT from
        this object — never present-and-empty, never sent for the client to
        hide (§11).
      properties:
        badges:
          type: array
          items:
            $ref: '#/components/schemas/ProfileBadge'
        degrees:
          type: array
          items:
            $ref: '#/components/schemas/ProfileDegree'
        courses:
          $ref: '#/components/schemas/ProfileCourses'
        activity_feed:
          type: array
          items:
            $ref: '#/components/schemas/ProfileActivityEvent'
        activity_heatmap:
          $ref: '#/components/schemas/ProfileHeatmap'
    Profile:
      type: object
      description: >-
        A learner's profile, already filtered to the viewer it was built for.
        There is no email field, for any viewer.
      properties:
        handle:
          type: string
        displayName:
          type:
            - string
            - "null"
          nullable: true
        bio:
          type:
            - string
            - "null"
          nullable: true
        joinedAt:
          type: string
          format: date-time
        avatar:
          $ref: '#/components/schemas/ProfileAvatar'
        noindex:
          type: boolean
        viewer:
          $ref: '#/components/schemas/ProfileViewer'
        sections:
          $ref: '#/components/schemas/ProfileSections'
        visibility:
          $ref: '#/components/schemas/ProfileVisibility'
      required:
        - handle
        - displayName
        - bio
        - joinedAt
        - avatar
        - noindex
        - viewer
        - sections
...